[chore] Chance the name of test file for LivingSpace
